refactor: field data interface
