+++ libs/backend/application/field/src/field.application.controller.ts
@@ -2,6 +2,6 @@ import { Body, Controller, Post } from '@nestjs/common'
 
 @Controller('field')
 export class FieldApplicationController {
-  @Post('create-field')
-  async createField(@Body() createFieldDto: ICreateFieldData) {}
+  // @Post('create-field')
+  // async createField(@Body() createFieldDto: ICreateFieldData) {}
 }
