+++ libs/frontend/abstract/domain/src/field/field.model.interface.ts
@@ -3,19 +3,16 @@ import type {
   FieldDeleteInput,
   FieldUpdateInput,
 } from '@codelab/shared/abstract/codegen'
-import type { IField, IFieldDto } from '@codelab/shared/abstract/core'
+import type {
+  IField,
+  IFieldDefaultValue,
+  IFieldDto,
+  IValidationRules,
+} from '@codelab/shared/abstract/core'
 import type { Nullish } from '@codelab/shared/abstract/types'
 import type { Ref } from 'mobx-keystone'
 import type { ICacheService, IModel } from '../shared'
 import type { IInterfaceTypeModel, ITypeModel } from '../type'
-import type { IValidationRules } from './field.validation'
-
-export type IFieldDefaultValue =
-  | Array<IFieldDefaultValue>
-  | boolean
-  | number
-  | string
-  | { [x: string]: IFieldDefaultValue }
 
 export interface IFieldModel<T extends ITypeModel = ITypeModel>
   extends Omit<
@@ -43,5 +40,3 @@ export interface IFieldModel<T extends ITypeModel = ITypeModel>
   detachPrevSibling(): void
   toUpdateNodesInput(): Pick<FieldUpdateInput, 'nextSibling' | 'prevSibling'>
 }
-
-export type IFieldRef = string
