+++ libs/frontend/abstract/domain/src/field/index.ts
@@ -1,8 +1,6 @@
-export * from './field.data.interface'
 export * from './field.domain.service.context'
 export * from './field.domain.service.interface'
 export * from './field.model.interface'
 export * from './field.node.interface'
 export * from './field.record.interface'
 export * from './field.ref'
-export * from './field.validation'
