+++ libs/frontend/application/type/src/interface-form/fields/select-default-value/SelectDefaultValue.tsx
@@ -1,12 +1,9 @@
-import type {
-  IPrimitiveTypeModel,
-  IValidationRules,
-} from '@codelab/frontend/abstract/domain'
+import type { IPrimitiveTypeModel } from '@codelab/frontend/abstract/domain'
 import { MODEL_ACTION } from '@codelab/frontend/abstract/types'
 import { useStore } from '@codelab/frontend/application/shared/store'
 import { Form } from '@codelab/frontend/presentation/view'
 import { PrimitiveTypeKind } from '@codelab/shared/abstract/codegen'
-import type { IPropData } from '@codelab/shared/abstract/core'
+import type { IPropData, IValidationRules } from '@codelab/shared/abstract/core'
 import { ITypeKind } from '@codelab/shared/abstract/core'
 import type { Maybe } from '@codelab/shared/abstract/types'
 import { useAsync, useMountEffect } from '@react-hookz/web'
