+++ libs/frontend/application/type/src/use-cases/fields/get-fields/validation.ts
@@ -1,7 +1,5 @@
-import type {
-  IValidationRules,
-  ValidationRuleTag,
-} from '@codelab/frontend/abstract/domain'
+import type { ValidationRuleTag } from '@codelab/frontend/abstract/domain'
+import type { IValidationRules } from '@codelab/shared/abstract/core'
 import type { Nullish } from '@codelab/shared/abstract/types'
 
 export const getValidationRuleTagsArray = (
