+++ libs/shared/abstract/core/src/index.ts
@@ -9,7 +9,7 @@ export * from './component'
 export * from './domain/domain.dto.interface'
 export * from './element'
 export * from './errors/domain'
-export * from './field/field.dto.interface'
+export * from './field'
 export * from './model/node-type.interface'
 export * from './page'
 export * from './prop/prop.dto.interface'
@@ -17,6 +17,6 @@ export * from './redirect'
 export * from './resource'
 export * from './stage.enum'
 export * from './store'
-export * from './tag/tag.dto.interface'
+export * from './tag'
 export * from './type'
 export * from './user'
