+++ libs/shared/abstract/core/src/tag/index.ts
@@ -1,0 +1,1 @@
+export * from './tag.dto.interface'
